@@ -1,26 +1,27 @@
-/*-
- * ============LICENSE_START=======================================================
- * ONAP
- * ================================================================================
- * Copyright (C) 2021 Wipro Limited.
- * ================================================================================
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+/*
+ *  ============LICENSE_START=======================================================
+ *  Copyright (C) 2021 Wipro Limited.
+ *  Modifications Copyright (C) 2023 Nordix Foundation
+ *  ================================================================================
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *        http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * ============LICENSE_END=========================================================
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *  ============LICENSE_END=========================================================
  */
 
 package org.onap.cps.tbdmt.rest;
 
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 import org.onap.cps.tbdmt.model.ExecutionRequest;
 import org.onap.cps.tbdmt.service.ExecutionBusinessLogic;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -47,8 +48,8 @@ public class ExecutionController {
      */
     @PostMapping(path = "/execute/{model}/{templateId}")
     public ResponseEntity<String> executeTemplate(@Valid @PathVariable final String model,
-        @Valid @PathVariable final String templateId,
-        @Valid @RequestBody final ExecutionRequest executionRequest) {
+                                                  @Valid @PathVariable final String templateId,
+                                                  @Valid @RequestBody final ExecutionRequest executionRequest) {
         final String result = executionBusinessLogic.executeTemplate(model, templateId, executionRequest);
         return new ResponseEntity<>(result, HttpStatus.OK);
     }
